+++ functions/src/utils/algolia.function.ts
